compiler: fix object allocation (experimental)
When an object's member (a slot value) is dynamically
allocated, then it must itself be dynamically allocated
so that the garbage collector keeps the member alive.

I modified the compiler macro (dhm-t) for 'new-empty'
to have it trigger dynamic object allocation. But not
sure this is the best way to do it, hence 'experimental'.



git-svn-id: https://lush.svn.sourceforge.net/svnroot/lush/lush2@651 b68d3702-14d2-4b18-b47c-39ed2627325e