changes to BLAS API: destructive functions return nil
Destructive BLAS-functions used to return the
updated argument. The problem with that is that
the compiler generates potentially problematic
code for those functions:

When the arguments to the destructive BLAS 
function are created in the same function but
not returned by that function, the allocation
would nevertheless be dynamic. This is bad when
a BLAS function is called in a loop, for instance.

The right thing to do would be to make the compiler
smarter, but I don't have the time to look into it.



git-svn-id: https://lush.svn.sourceforge.net/svnroot/lush/lush2@688 b68d3702-14d2-4b18-b47c-39ed2627325e