followup on r684 (fix for allocation in a loop problem)
This patch fixes the 'idx-bloop' compiler macro.

The fix is similar to the one for the 'for' macro,
but this time the number of symbols in :treetype:symbol
that belong outside the loop body is unknown. Therefore
'dhc-push-loop-scope-t' tags the (beginning of the)
symbols that belong inside he loop body, and the 
dhm-c for 'idx-bloop' looks for the tag to split the
list :treetype:symbol accordingly. 

This is just another ugly hack, of course. But I have
pretty much given up on the compiler. It should be 
rewritten from scratch.




git-svn-id: https://lush.svn.sourceforge.net/svnroot/lush/lush2@723 b68d3702-14d2-4b18-b47c-39ed2627325e